tests: detailsのテスト分割
