+++ src/features/editor/extensions/nodes/details/tests/key.browser.spec.ts
@@ -1,87 +1,13 @@
 import Document from "@tiptap/extension-document";
 import Paragraph from "@tiptap/extension-paragraph";
 import Text from "@tiptap/extension-text";
-import { page, userEvent } from "@vitest/browser/context";
+import { userEvent } from "@vitest/browser/context";
 import { describe, expect, it } from "vitest";
 import { waitSelectionChange } from "@/tests/dom";
 import { renderTiptapEditor } from "@/tests/editor";
-import { DetailsContent } from "./content";
-import { Details } from "./index";
-import { DetailsSummary } from "./summary";
-
-describe("InputRule", () => {
-  it(":::details で アコーディオンブロックが作成される", async () => {
-    const editor = renderTiptapEditor({
-      content: "<p>Text</p>",
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().run();
-    });
-    await userEvent.keyboard(":::details ");
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(
-      'doc(details(detailsSummary, detailsContent(paragraph("Text"))))',
-    );
-    expect(editor.state.selection.from).toBe(2);
-  });
-
-  it("行の途中では InputRule が発動しない", async () => {
-    const editor = renderTiptapEditor({
-      content: "<p>Text</p>",
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(2).run();
-    });
-    await userEvent.keyboard(":::details ");
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe('doc(paragraph("T:::details ext"))');
-  });
-
-  it("アコーディオンのタイトル内では InputRule が発動しない", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details><summary>Title</summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor.chain().focus().setTextSelection(2).run();
-    });
-    await userEvent.keyboard(":::details ");
-
-    const docString = editor.state.doc.toString();
-    expect(docString).toBe(
-      'doc(details(detailsSummary(":::details Title"), detailsContent(paragraph("Text"))))',
-    );
-  });
-});
+import { Details } from "../";
+import { DetailsContent } from "../content";
+import { DetailsSummary } from "../summary";
 
 describe("キー入力", () => {
   describe("Backspace", () => {
@@ -306,114 +232,5 @@ describe("キー入力", () => {
 
       expect(editor.state.selection.from).toBe(13); // "After" の最初
     });
-  });
-});
-
-describe("範囲選択", () => {
-  it("前のノードとコンテンツの間で範囲選択してBackspaceを押すと、アコーディオンが削除される", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<p>Before</p><details open="true"><summary></summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor
-        .chain()
-        .focus()
-        .setTextSelection({
-          from: 5,
-          to: 14,
-        })
-        .run();
-    });
-
-    await userEvent.keyboard("{Backspace}");
-
-    expect(editor.state.doc.toString()).toBe('doc(paragraph("Befoext"))');
-    expect(editor.state.selection.from).toBe(5);
-  });
-
-  it("アコーディオンと次のノードの間で範囲選択してBackspaceを押すと、アコーディオンのコンテンツに結合する", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details open><summary>Title</summary><div class="details-content"><p>Text</p></div></details><p>After</p>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    await waitSelectionChange(() => {
-      editor
-        .chain()
-        .focus()
-        .setTextSelection({
-          from: 11,
-          to: 19,
-        })
-        .run();
-    });
-    await userEvent.keyboard("{Backspace}");
-
-    expect(editor.state.doc.toString()).toBe(
-      'doc(details(detailsSummary("Title"), detailsContent(paragraph("Tfter"))))',
-    );
-    expect(editor.state.selection.from).toBe(11);
-  });
-});
-
-describe("アコーディオンの開閉", () => {
-  it("ボタンを押して閉じる", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details open><summary>Title</summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    const button = page.getByRole("button");
-    await userEvent.click(button);
-
-    const detailsNode = editor.state.doc.firstChild;
-    expect(detailsNode?.attrs.open).toBe(false);
-  });
-
-  it("ボタンを押して開く", async () => {
-    const editor = renderTiptapEditor({
-      content:
-        '<details><summary>Title</summary><div class="details-content"><p>Text</p></div></details>',
-      extensions: [
-        Document,
-        Paragraph,
-        Text,
-        Details,
-        DetailsContent,
-        DetailsSummary,
-      ],
-    });
-
-    const button = page.getByRole("button");
-    await userEvent.click(button);
-
-    const detailsNode = editor.state.doc.firstChild;
-    expect(detailsNode?.attrs.open).toBe(true);
   });
 });
