+++ src/features/editor/extensions/nodes/details/tests/render.spec.ts
@@ -1,0 +1,33 @@
+import Document from "@tiptap/extension-document";
+import Paragraph from "@tiptap/extension-paragraph";
+import Text from "@tiptap/extension-text";
+import { describe, expect, it } from "vitest";
+import { createEditorInstance } from "@/tests/editor-instance";
+import { Details } from "..";
+import { DetailsContent } from "../content";
+import { DetailsSummary } from "../summary";
+
+describe("HTMLのパース・レンダリング", () => {
+  it("アコーディオンが正しいHTMLでレンダリングされる", () => {
+    const editor = createEditorInstance({
+      extensions: [
+        Document,
+        Paragraph,
+        Text,
+        Details,
+        DetailsContent,
+        DetailsSummary,
+      ],
+      content:
+        '<details><summary></summary><div class="details-content"><p>テキスト</p></div></details>',
+    });
+
+    const html = editor.getHTML();
+    expect(html).toContain("<details>");
+    expect(html).toContain("<summary></summary>");
+    expect(html).toContain('<div class="details-content">');
+    expect(html).toContain("<p>テキスト</p>");
+    expect(html).toContain("</div>");
+    expect(html).toContain("</details>");
+  });
+});
